requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-storage
feedparser
pandas